python/uagents-adapter/pyproject.toml
[project]
name = "uagents-adapter"
version = "0.5.0"
description = "Adapters for uAgents to integrate with LangChain, CrewAI, MCP, and A2A"
authors = [{ name = "Fetch.AI Limited" }]
license = { text = "Apache 2.0" }
readme = "README.md"
keywords = ["uagents", "agents", "langchain", "crewai", "mcp", "a2a", "fetch.ai"]
requires-python = ">=3.10,<4.0"
dependencies = [
    "uagents (>=0.22.3)",
    "pydantic (>=2.8,<3.0)",
    "langchain-core (>=0.3.52,<0.4.0)",
    "requests (>=2.32.0,<3.0)",
]

[project.optional-dependencies]
mcp = ["mcp (>=1.8.1)"]
a2a-inbound = ["a2a-sdk[all,sqlite]>=0.2.11", "uvicorn>=0.27.0", "httpx>=0.25.0", "click>=8.0.0", "python-dotenv>=1.0.0"]

[tool.poetry]
# install from src/uagents_adapter folder
packages = [{ include = "uagents_adapter", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
ruff = "^0.11.0"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    # pycodestyle (Errors, Warnings)
    "E",
    "W",
    # Pyflakes
    "F",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # pep8-naming
    "N",
    # pylint
    "PL",
]
ignore = ["PLR0913", "PLR0912", "PLR0911", "PLR2004", "PLR0915"]

[tool.ruff.lint.pycodestyle]
max-line-length = 100 